Add a single, cleaner entrypoint
I don't know why, but I could not get this working with npm's
concurrently. There's something different between the way that node and
python spawn subprocesses. I'd like to look into this more when I have
some time, but for now, I've got a quick python script to handle this.